refactor(daemon): extract transport auth + helpers into a transports module (#279 item 4)
main.rs was a ~2.4k-line wiring god-function. Extract the self-contained
transport-setup pieces into a new `transports` module:

- The four WS/UDS auth validators (`WsSettingsAuth`, `OidcAwareAuth`,
  `WsAuthDiscoveryProvider`, `WsAsUdsAuth`) and the `WsBasicLogin` login
  service + `WsLoginMode`.
- The transport enable/disable defaults (`transport_defaults`) and the
  env/host resolution helpers (`resolve_uds_socket_path`, `env_bool`/
  `parse_env_bool`, `daemon_host_label`, `is_container_environment`,
  `resolve_ws_login_mode`/`_decision`).
- Their unit tests move with them (now `transports::tests`).

Types are `pub(crate)`; main.rs wires them via a single `use`. Pure code
movement — no behavior change. main.rs drops ~700 lines.
